perf(train,model): fused RMSNorm, compiled CE+RoPE, foreach_norm, inference_opt fixes
Training speedup ~2.8x (2981ms -> 1063ms/step), +180% tokens/sec.
VRAM reduction ~34% (8.33GB -> 5.49GB).

Changes:
- transformer.py: fused torch.nn.functional.rms_norm, apply_rope via cat,
  conditional torch.compile for CUDA, forward_repacked compile path,
  SwiGLU/GQA state_dict load fix for fused weights
- losses.py: conditional torch.compile for cross_entropy_loss on CUDA
- optim.py: get_grad_norm via torch._foreach_norm
- moe.py: dtype fix for index_copy_/scatter_add_ with manual_bf16
- inference_opt.py: fold_rmsnorm + repack_qkv adapted to fused weights
- train.py: get_grad_norm foreach optimization (reverted)
- tests: updated test_config, test_configs, test_muon_optim for reality

Constraint: torch.compile on Windows requires MSVC; conditional on CUDA only
Rejected: foreach clip_grad_norm (memory spike)
Confidence: high
Scope-risk: moderate (state_dict format changed)
Not-tested: full training run on real data beyond 10 steps

+++ src/hagi/model/moe.py
@@ -72,10 +72,10 @@ class MoESwiGLU(nn.Module):
                 tokens = flat.index_select(0, indices)
                 expert_out = expert(tokens)
                 if self.top_k == 1:
-                    output.index_copy_(0, indices, expert_out)
+                    output.index_copy_(0, indices, expert_out.to(output.dtype))
                 else:
                     idx = indices.unsqueeze(-1).expand(-1, expert_out.size(-1))
-                    output.scatter_add_(0, idx, expert_out * probs.index_select(0, indices).unsqueeze(-1))
+                    output.scatter_add_(0, idx, expert_out.to(output.dtype) * probs.index_select(0, indices).unsqueeze(-1))
 
         output = output.view(B, T, D)
